move btl functions out of ctd module
